requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
albucore==0.0.15
albumentations==1.4.14
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
bidict==0.23.1
bitsandbytes==0.43.3
blake3==0.4.1
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clip-anytorch==2.6.0
colorama==0.4.6
coloredlogs==15.0.1
compel==2.0.2
contourpy==1.3.0
controlnet_aux==0.0.7
cycler==0.12.1
datasets==2.19.1
Deprecated==1.2.14
diffusers==0.30.2
dill==0.3.8
dnspython==2.4.2
dynamicprompts==0.31.0
easing-functions==1.0.4
einops==0.8.0
email_validator==2.2.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
facexlib==0.3.0
fastapi==0.111.0
fastapi-cli==0.0.5
fastapi-events==0.11.1
filelock==3.15.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.3
future==1.0.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.7
imageio==2.35.0
importlib_metadata==7.1.0
intel-openmp==2021.4.0
invisible-watermark==0.2.0
InvokeAI==4.2.9
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.7
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.7
mkl==2021.4.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
npyscreen==4.10.5
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.16.3
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
peft==0.11.1
picklescan==0.0.16
pillow==10.3.0
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.8
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.7.2
pydantic-settings==2.2.1
pydantic_core==2.18.3
Pygments==2.18.0
Pympler==1.0.1
pyparsing==3.1.4
PyPatchMatch==1.0.1
pyperclip==1.9.0
pyreadline3==3.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.1
pytorch-lightning==2.1.3
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.28.2
rich==13.8.1
safetensors==0.4.3
scikit-image==0.21.0
scipy==1.14.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
spandrel==0.3.4
starlette==0.37.2
sympy==1.12.1
tbb==2021.12.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
test_tube==0.7.5
tifffile==2024.8.10
timm==0.6.13
tokenizers==0.19.1
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchmetrics==0.11.4
torchsde==0.2.6
torchvision==0.19.1+cu124
tqdm==4.66.4
trampoline==0.1.2
transformers==4.41.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.28.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
Werkzeug==3.0.4
windows-curses==2.3.3
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.11.1
zipp==3.19.2
